+++ R/vosonSML-package.R
@@ -36,7 +36,7 @@
 #' @importFrom magrittr '%>%' '%<>%'
 #' @importFrom dplyr rename group_by summarise ungroup left_join select mutate filter coalesce row_number
 #' distinct anti_join mutate_all mutate_at ends_with vars funs bind_rows arrange case_when if_else
-#' contains
+#' contains n rowwise
 #' @importFrom rlang '.data'
 #' @importFrom textutils HTMLdecode
 #' @importFrom tictoc tic toc